[core] Move type traits to dedicated file
